[#138] Move jumbotron into components
