+++ frontend/src/pages/landing/components/index.js
@@ -1,3 +1,4 @@
 export { default as GetStarted } from "./GetStarted";
 export { default as ExploreStudies } from "./ExploreStudies";
 export { default as FooterDisclaimer } from "./FooterDisclaimer";
+export { default as Jumbotron } from "./Jumbotron";
