+++ frontend/src/pages/landing/components/landingcomp.scss
@@ -1,12 +1,40 @@
 @import "../../../variables.scss";
 
+#jumbotron {
+  background-image: url("/assets/images/jumbotron.jpeg") !important;
+  background-size: cover;
+  background-repeat: no-repeat;
+  background-position: center -7rem;
+  width: 100%;
+  height: 43rem;
+  padding: 4rem $default-padding-size $default-padding-size
+    $default-padding-size;
+
+  h1 {
+    font-family: "TabletGothicBold" !important;
+    color: #ffffff;
+    font-weight: 700;
+    font-size: 3.4rem !important;
+    width: 606px;
+    line-height: 70px;
+  }
+
+  h3 {
+    color: #ffffff;
+    font-weight: 400;
+    font-size: 1.9rem;
+    line-height: 45px;
+    width: 426px;
+  }
+}
+
 #get-started {
   padding: 3rem $default-padding-size;
   background: #fcf9f4;
 
   h2 {
     color: #00403e;
-    font-family: "Roc Grotesk";
+    font-family: "RocGrotesk";
     font-size: 30px;
     font-style: normal;
     font-weight: 700;
