+++ frontend/src/pages/landing/landing.scss
@@ -2,32 +2,6 @@
 
 #landing,
 .landing-container {
-  .jumbotron-wrapper {
-    background-image: url("/assets/images/jumbotron.jpeg") !important;
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position: center -7rem;
-    width: 100%;
-    height: 43rem;
-    padding: 4rem $default-padding-size $default-padding-size
-      $default-padding-size;
-
-    h1 {
-      font-family: "TabletGothicBold" !important;
-      color: #ffffff;
-      font-weight: 700;
-      font-size: 3.4rem !important;
-      width: 606px;
-    }
-
-    h3 {
-      color: #ffffff;
-      font-weight: 400;
-      font-size: 1.9rem;
-      width: 426px;
-    }
-  }
-
   .income-driver-framework-wrapper {
     background-color: #f3f7f6;
     width: 100%;
